+++ cmd/main.go
@@ -1,26 +1,32 @@
 package main
 
 import (
-  logger "github.com/sirupsen/logrus"
-  "os"
-  "os/signal"
-  "syscall"
+	"os"
+	"os/signal"
+	"syscall"
+
+	logger "github.com/sirupsen/logrus"
 )
 
 func main() {
-  logger.Infof("Starting ms-mail-sender")
+	logger.Infof("Starting ms-mail-sender")
 
-  signalChan := make(chan os.Signal, 1)
-  signal.Notify(signalChan, os.Interrupt, syscall.SIGTERM)
-  defer handlePanic()
+	signalChan := make(chan os.Signal, 1)
+	signal.Notify(signalChan, os.Interrupt, syscall.SIGTERM)
+	app := injectApps()
+	defer handlePanic()
 
-  <-signalChan
+	for _, a := range app {
+		a.RunConsumers()
+	}
 
-  logger.Infof("Shutting down ms-mail-sender")
+	<-signalChan
+
+	logger.Infof("Shutting down ms-mail-sender")
 }
 
 func handlePanic() {
-  if r := recover(); r != nil {
-    logger.WithField("panic", r).Error("Panic occurred in the application.")
-  }
+	if r := recover(); r != nil {
+		logger.WithField("panic", r).Error("Panic occurred in the application.")
+	}
 }