+++ internal/mail/domain/commands/container.go
@@ -1,0 +1,9 @@
+package commands
+
+import "github.com/google/wire"
+
+//nolint:gochecknoglobals // requirement for container
+var Container = wire.NewSet(
+	NewSendMailCommand,
+	wire.Bind(new(SendMail), new(*SendMailCommand)),
+)