+++ internal/mail/domain/entities/email_type.go
@@ -1,0 +1,8 @@
+package entities
+
+type EmailType int
+
+const (
+	Success = iota
+	Fail
+)