+++ internal/mail/infrastructure/listeners/container.go
@@ -1,0 +1,9 @@
+package listeners
+
+import "github.com/google/wire"
+
+//nolint:gochecknoglobals // requirement for container
+var Container = wire.NewSet(
+	NewSendSuccessMailQueueListener,
+	NewSendErrorMailQueueListener,
+)